start separation of game files into scenes. Comment game logic for now.
